Add files via upload
In this project walkthrough, we'll create a movie recommendation system using Jupyter, Python, and Pandas.  In this we are able to type the name of a movie into an input box, and instantly get recommendations for other movies we might like.

We'll start with the  Dataset for Movies, which contains movie reviews and ratings.  Then, we'll build a search engine to find a specific movie title in our data.  We'll then be able to create a recommendation engine to recommend specific movies.